Use chunkalyse to support multiple entry points
