--- lib/flatten/index.js
@@ -1,1 +1,0 @@
-module.exports = items => items.every(Array.isArray) ? flatten([].concat(...items)) : items;
