--- programs/diff-summary/bundles/entry/index.js
@@ -1,3 +1,0 @@
-const byteSize = require('byte-size');
-
-module.exports = ({size, percent}) => size === 0 ? '0' : `${byteSize(size)} (${percent}%)`;
